requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
django-rest==0.8.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
flake8==6.0.0
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
yarl==1.8.2
